AG - Moved driver record trip expenses tests to own class
